# backend/internal/docs/swagger.yaml
basePath: /
definitions:
  github_com_web-kokhi_backend_internal_forecast.Point:
    properties:
      aqi:
        type: integer
      forecast:
        type: boolean
      pm25:
        type: number
      time:
        type: string
    type: object
  github_com_web-kokhi_backend_internal_forecast.Result:
    properties:
      current_pm25:
        type: number
      health_advice:
        type: string
      hours_until_warning:
        type: number
      message:
        type: string
      points:
        items:
          $ref: '#/definitions/github_com_web-kokhi_backend_internal_forecast.Point'
        type: array
      risk_level:
        type: string
      slope_per_hour:
        type: number
      station_id:
        type: integer
      station_name:
        type: string
      sufficient_data:
        type: boolean
      trend:
        type: string
      will_exceed_warning:
        type: boolean
    type: object
  github_com_web-kokhi_backend_internal_model.Alert:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      metric:
        type: string
      station_id:
        type: integer
      station_name:
        type: string
      value:
        type: number
    type: object
  github_com_web-kokhi_backend_internal_model.Gauge:
    properties:
      max:
        type: number
      type:
        type: string
      value:
        type: number
    type: object
  github_com_web-kokhi_backend_internal_model.HeatmapPoint:
    properties:
      aqi:
        type: integer
      lat:
        type: number
      lng:
        type: number
      pm25:
        type: number
      recorded_at:
        type: string
      station_id:
        type: integer
      station_name:
        type: string
    type: object
  github_com_web-kokhi_backend_internal_model.HistoryPoint:
    properties:
      pm10:
        type: number
      pm25:
        type: number
      time:
        type: string
    type: object
  github_com_web-kokhi_backend_internal_model.ReadingInput:
    properties:
      firmware:
        type: string
      humidity:
        type: number
      pm10:
        type: number
      pm25:
        type: number
      rssi:
        type: integer
      station_id:
        type: integer
      temp:
        type: number
    type: object
  github_com_web-kokhi_backend_internal_model.ReadingLog:
    properties:
      aqi:
        type: integer
      humidity:
        type: number
      id:
        type: integer
      pm10:
        type: number
      pm25:
        type: number
      quality_flag:
        type: string
      recorded_at:
        type: string
      station_id:
        type: integer
      station_name:
        type: string
      temp:
        type: number
    type: object
  github_com_web-kokhi_backend_internal_model.ReadingStats:
    properties:
      aqi_avg:
        type: integer
      aqi_max:
        type: integer
      aqi_min:
        type: integer
      count:
        type: integer
      humidity_avg:
        type: number
      humidity_max:
        type: number
      humidity_min:
        type: number
      pm10_avg:
        type: number
      pm10_max:
        type: number
      pm10_min:
        type: number
      pm25_avg:
        type: number
      pm25_max:
        type: number
      pm25_min:
        type: number
      station_id:
        type: integer
      station_name:
        type: string
      temp_avg:
        type: number
      temp_max:
        type: number
      temp_min:
        type: number
    type: object
  github_com_web-kokhi_backend_internal_model.Settings:
    properties:
      aqi_critical:
        type: integer
      aqi_warning:
        type: integer
      pm25_critical:
        type: number
      pm25_warning:
        type: number
      telegram_chat_id:
        type: string
      telegram_enabled:
        type: boolean
      telegram_ready:
        type: boolean
    type: object
  github_com_web-kokhi_backend_internal_model.Station:
    properties:
      aqi:
        type: integer
      aqiLabel:
        type: string
      color:
        type: string
      gauges:
        items:
          $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Gauge'
        type: array
      humidity:
        type: number
      id:
        type: integer
      last_seen:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      online:
        type: boolean
      pm25:
        type: number
      subtitle:
        type: string
      temp:
        type: number
      time:
        type: string
    type: object
  github_com_web-kokhi_backend_internal_model.StationStatus:
    properties:
      id:
        type: integer
      last_aqi:
        type: integer
      last_pm25:
        type: number
      last_seen:
        type: string
      name:
        type: string
      online:
        type: boolean
    type: object
  github_com_web-kokhi_backend_internal_model.SystemStatus:
    properties:
      alerts_today:
        type: integer
      api_status:
        type: string
      db_status:
        type: string
      mode:
        type: string
      readings_today:
        type: integer
      stale_threshold_sec:
        type: integer
      started_at:
        type: string
      stations:
        items:
          $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.StationStatus'
        type: array
      stations_online:
        type: integer
      stations_total:
        type: integer
      uptime_sec:
        type: integer
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        example: invalid JSON body
        type: string
    type: object
  internal_handler.HealthResponse:
    properties:
      db:
        example: postgres
        type: string
      mode:
        example: simulator
        type: string
      status:
        example: ok
        type: string
    type: object
  internal_handler.TelegramTestRequest:
    properties:
      chat_id:
        example: "123456789"
        type: string
    type: object
  internal_handler.TelegramTestResponse:
    properties:
      status:
        example: sent
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API giám sát chất lượng không khí — trạm đo, cảnh báo, dự báo và thống
    kê.
  title: Web Kokhi API
  version: "1.0"
paths:
  /api/alerts:
    get:
      parameters:
      - description: Lọc theo trạm
        in: query
        name: station_id
        type: integer
      - default: 24
        description: Số giờ lùi lại
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Alert'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Danh sách cảnh báo
      tags:
      - alerts
  /api/forecast:
    get:
      description: Dự báo theo hồi quy tuyến tính; bao gồm health_advice và risk_level
      parameters:
      - default: 6
        description: Số bước dự báo
        in: query
        name: steps
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_forecast.Result'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Dự báo PM2.5 / AQI
      tags:
      - analytics
  /api/heatmap:
    get:
      description: Tọa độ và PM2.5/AQI theo thời gian cho bản đồ nhiệt
      parameters:
      - default: 6
        description: Số giờ lùi lại
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.HeatmapPoint'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Điểm heatmap
      tags:
      - analytics
  /api/readings:
    get:
      parameters:
      - description: Lọc theo trạm
        in: query
        name: station_id
        type: integer
      - default: 24
        description: Số giờ lùi lại
        in: query
        name: hours
        type: integer
      - default: 200
        description: Giới hạn số dòng
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.ReadingLog'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Nhật ký bản ghi đo
      tags:
      - readings
    post:
      consumes:
      - application/json
      description: Nhận bản ghi PM2.5/PM10 từ thiết bị. Yêu cầu header X-Station-Key
        khi STATION_API_KEY được cấu hình.
      parameters:
      - description: API key trạm (bắt buộc nếu server có STATION_API_KEY)
        in: header
        name: X-Station-Key
        type: string
      - description: Dữ liệu đo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.ReadingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Station'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - StationKey: []
      summary: Ghi nhận dữ liệu cảm biến (ESP32)
      tags:
      - readings
  /api/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Settings'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Lấy cấu hình cảnh báo
      tags:
      - settings
    put:
      consumes:
      - application/json
      parameters:
      - description: Ngưỡng PM2.5/AQI và Telegram
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Settings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Cập nhật cấu hình cảnh báo
      tags:
      - settings
  /api/stations:
    get:
      description: Trả về tất cả trạm với chỉ số AQI, PM2.5 và gauge mới nhất
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Station'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Danh sách trạm
      tags:
      - stations
  /api/stations/{id}:
    get:
      parameters:
      - description: Station ID (1-3)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.Station'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Chi tiết một trạm
      tags:
      - stations
  /api/stations/{id}/history:
    get:
      parameters:
      - description: Station ID (1-3)
        in: path
        name: id
        required: true
        type: integer
      - default: 24
        description: Số giờ lùi lại
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.HistoryPoint'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Lịch sử đo của trạm
      tags:
      - stations
  /api/stats/readings:
    get:
      parameters:
      - default: 24
        description: Số giờ lùi lại
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.ReadingStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Thống kê bản ghi đo
      tags:
      - analytics
  /api/system/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_web-kokhi_backend_internal_model.SystemStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Trạng thái hệ thống
      tags:
      - system
  /api/telegram/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: Chat ID (tùy chọn, mặc định dùng cấu hình)
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.TelegramTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.TelegramTestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Gửi tin nhắn Telegram thử
      tags:
      - settings
  /health:
    get:
      description: Kiểm tra trạng thái API và kết nối PostgreSQL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.HealthResponse'
      summary: Health check
      tags:
      - system
  /ws/stations:
    get:
      description: 'Kết nối WebSocket tại ws://host/ws/stations. Server broadcast
        JSON: `{"type":"stations","data":[...]}`'
      responses: {}
      summary: WebSocket cập nhật trạm realtime
      tags:
      - realtime
schemes:
- http
securityDefinitions:
  StationKey:
    description: API key cho POST /api/readings (khi biến môi trường STATION_API_KEY
      được đặt)
    in: header
    name: X-Station-Key
    type: apiKey
swagger: "2.0"